💻 Refactor retrace state

+++ src/enemies.rs
@@ -1,22 +1,13 @@
 use bevy::prelude::*;
 
 use crate::force::Force;
-use crate::grid::{get_coordinates, get_indeces};
 use crate::pathfinding::VectorField;
-use crate::physics::Moving;
-
-#[derive(Default, Debug, PartialEq, Eq)]
-pub enum EnemyState {
-    #[default]
-    WalkToTower,
-    WalkBack,
-}
+use crate::physics::{Collider, Moving, MovingState};
 
 #[derive(Component, Default)]
 pub struct Enemy {
     /// Tracks when enemy should turn around and leave.
     pub timer: Timer,
-    pub state: EnemyState,
 }
 
 pub struct EnemyPlugin;
@@ -29,44 +20,27 @@ impl Plugin for EnemyPlugin {
 
 fn turn_enemy(
     mut force_query: Query<(&Force, &Transform)>,
-    mut enemy_query: Query<(&Enemy, &mut Moving, &Transform)>,
+    mut enemy_query: Query<(&mut Moving, &Transform), With<Enemy>>,
     vector_field: Res<VectorField>,
     time: Res<Time>,
 ) {
-    for (enemy, mut moving, transform) in &mut enemy_query {
+    for (mut moving, transform) in &mut enemy_query {
+        // Skip when retracing
+        if moving.state == MovingState::Retrace {
+            continue;
+        }
+
         // Slowly point enemy towards tower
-        let mut force_sum = match enemy.state {
-            EnemyState::WalkToTower => {
-                vector_field.get_direction(transform.translation.truncate()) * moving.speed.abs()
-            }
-            EnemyState::WalkBack => {
-                // Remove last indeces from route history if we've arrived there.
-                if let Some(last_indeces) = moving.route_history.last() {
-                    let current_indeces = get_indeces(transform.translation.truncate());
-                    if current_indeces == *last_indeces {
-                        moving.route_history.pop();
-                    }
-                }
+        let mut force_sum =
+            vector_field.get_direction(transform.translation.truncate()) * moving.speed.abs();
 
-                match moving.route_history.last() {
-                    Some((x, y)) => {
-                        let current_coordinates = get_coordinates(*x, *y);
-                        current_coordinates - transform.translation.truncate()
-                    }
-                    None => Vec2::ZERO,
-                }
-            }
-        };
-
-        // Add external forces (except for when enemy walks back)
-        if enemy.state != EnemyState::WalkBack {
-            for (force, force_transform) in &mut force_query {
-                if let Some(f) = force.get_force(
-                    transform.translation.truncate(),
-                    force_transform.translation.truncate(),
-                ) {
-                    force_sum += f;
-                }
+        // Add external forces
+        for (force, force_transform) in &mut force_query {
+            if let Some(f) = force.get_force(
+                transform.translation.truncate(),
+                force_transform.translation.truncate(),
+            ) {
+                force_sum += f;
             }
         }
 
@@ -76,13 +50,18 @@ fn turn_enemy(
     }
 }
 
-fn update_timer(mut enemy_query: Query<&mut Enemy>, time: Res<Time>) {
-    for mut enemy in enemy_query.iter_mut() {
-        if enemy.state == EnemyState::WalkToTower {
+fn update_timer(
+    mut commands: Commands,
+    mut enemy_query: Query<(Entity, &mut Enemy, &mut Moving)>,
+    time: Res<Time>,
+) {
+    for (entity, mut enemy, mut moving) in enemy_query.iter_mut() {
+        if moving.state == MovingState::Normal {
             enemy.timer.tick(time.delta());
 
             if enemy.timer.finished() {
-                enemy.state = EnemyState::WalkBack;
+                moving.state = MovingState::Retrace;
+                commands.entity(entity).remove::<Collider>();
             }
         }
     }
